refactor : 불필요한 컨트롤러 없애기
