+++ src/main/java/com/opus/opus/modules/team/application/convenience/TeamConvenience.java
@@ -1,8 +1,10 @@
 package com.opus.opus.modules.team.application.convenience;
 
 import static com.opus.opus.modules.team.exception.TeamExceptionType.CONTEST_HAS_TEAM;
+import static com.opus.opus.modules.team.exception.TeamExceptionType.NOT_FOUND_TEAM;
 import static com.opus.opus.modules.team.exception.TeamExceptionType.TRACK_HAS_TEAM;
 
+import com.opus.opus.modules.team.domain.Team;
 import com.opus.opus.modules.team.domain.dao.TeamRepository;
 import com.opus.opus.modules.team.exception.TeamException;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +16,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class TeamConvenience {
     private final TeamRepository teamRepository;
+
+    public Team getValidateExistTeam(Long teamId) {
+        return teamRepository.findById(teamId)
+                .orElseThrow(() -> new TeamException(NOT_FOUND_TEAM));
+    }
 
     public void validateAllTeamsDeletedInContest(final Long contestId) {
         if (teamRepository.existsByContestId(contestId)) {
