+++ src/main/java/com/opus/opus/modules/team/application/dto/request/TeamContestAwardUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.opus.opus.modules.team.dto.request;
+package com.opus.opus.modules.team.application.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
